+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - [167](https://github.com/vegaprotocol/vegacapsule/issues/167) Fix validators filter in tendermint generator
 - [188](https://github.com/vegaprotocol/vegacapsule/issues/188) Support new changes for Ethereum RPC endpoint in Vega configuration
 - [209](https://github.com/vegaprotocol/vegacapsule/pull/209) Save tendermint template after merge to given file
+- [236](https://github.com/vegaprotocol/vegacapsule/pull/236) Improve detection for Nomad pending jobs